requirements.txt
streamlit
langchain
langchain-classic
langchain-huggingface
langchain-core
langchain-groq
python-dotenv